refactor: rename server env file
